requirements.txt
Flask==2.1.2
Flask-CORS==3.0.10
pandas==1.4.2
numpy==1.22.4
requests==2.28.1
SQLAlchemy==1.4.41
IMDbPY==6.
python-dotenv==0.21.0
werkzeug==2.0.3